verify OpenType files first with mkeot since it is faster than ttx

+++ src/bat/prerun.py
@@ -627,28 +627,33 @@ def verifyOTF(filename, tempdir=None, tags=[], offsets={}, debug=False, envvars=
 	if not lengthmatch:
 		return newtags
 
-	## first create a temporary directory where ttx can write its temporary files
-	fontdir = tempfile.mkdtemp(dir=unpacktempdir)
-	## now check if it is a valid file by running ogginfo
-	p = subprocess.Popen(['ttx', '-d', fontdir, '-i', filename], stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE, close_fds=True)
+	## run mkeot first. If it fails (mkeot might not be able to handle
+	## all OTF fonts) use ttx to dump fonts
+	p = subprocess.Popen(['mkeot', filename], stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE, close_fds=True)
 	(stanout, stanerr) = p.communicate()
 	if p.returncode != 0:
+		## first create a temporary directory where ttx can write its temporary files
+		fontdir = tempfile.mkdtemp(dir=unpacktempdir)
+		## now check if it is a valid file by running ttx
+		p = subprocess.Popen(['ttx', '-d', fontdir, '-i', filename], stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE, close_fds=True)
+		(stanout, stanerr) = p.communicate()
+		if p.returncode != 0:
+			## cleanup
+			## TODO: sanity checks
+			rmfiles = os.listdir(fontdir)
+			for r in rmfiles:
+				os.unlink(os.path.join(fontdir, r))
+			os.rmdir(fontdir)
+			return newtags
+		else:
+			## TODO: process output of ttx, since it might return 0 even though the font file is corrupted
+			pass
 		## cleanup
 		## TODO: sanity checks
 		rmfiles = os.listdir(fontdir)
 		for r in rmfiles:
 			os.unlink(os.path.join(fontdir, r))
 		os.rmdir(fontdir)
-		return newtags
-	else:
-		## TODO: process output of ttx, since it might return 0 even though the font file is corrupted
-		pass
-	## cleanup
-	## TODO: sanity checks
-	rmfiles = os.listdir(fontdir)
-	for r in rmfiles:
-		os.unlink(os.path.join(fontdir, r))
-	os.rmdir(fontdir)
 	newtags.append('otf')
 	newtags.append('font')
 	newtags.append('resource')
